Move simplify_requirement_list to native
